[libc][cleanup] Fix most conversion warnings
Automated commit created by applying diff 546640

Phabricator-ID: PHID-HMBT-fnfe6t6e2egn7nzjn6w3
Review-ID: D156950

+++ libc/src/__support/File/linux/file.cpp
@@ -44,7 +44,8 @@ namespace {
 
 FileIOResult write_func(File *f, const void *data, size_t size) {
   auto *lf = reinterpret_cast<LinuxFile *>(f);
-  int ret = __llvm_libc::syscall_impl(SYS_write, lf->get_fd(), data, size);
+  int ret = static_cast<int>(
+      __llvm_libc::syscall_impl(SYS_write, lf->get_fd(), data, size));
   if (ret < 0) {
     return {0, -ret};
   }
@@ -53,7 +54,8 @@ FileIOResult write_func(File *f, const void *data, size_t size) {
 
 FileIOResult read_func(File *f, void *buf, size_t size) {
   auto *lf = reinterpret_cast<LinuxFile *>(f);
-  int ret = __llvm_libc::syscall_impl(SYS_read, lf->get_fd(), buf, size);
+  int ret = static_cast<int>(
+      __llvm_libc::syscall_impl(SYS_read, lf->get_fd(), buf, size));
   if (ret < 0) {
     return {0, -ret};
   }
@@ -64,12 +66,13 @@ ErrorOr<long> seek_func(File *f, long offset, int whence) {
   auto *lf = reinterpret_cast<LinuxFile *>(f);
   long result;
 #ifdef SYS_lseek
-  int ret = __llvm_libc::syscall_impl(SYS_lseek, lf->get_fd(), offset, whence);
+  int ret = static_cast<int>(
+      __llvm_libc::syscall_impl(SYS_lseek, lf->get_fd(), offset, whence));
   result = ret;
 #elif defined(SYS__llseek)
   long result;
-  int ret = __llvm_libc::syscall_impl(SYS__llseek, lf->get_fd(), offset >> 32,
-                                      offset, &result, whence);
+  int ret = static_cast<int>(__llvm_libc::syscall_impl(
+      SYS__llseek, lf->get_fd(), offset >> 32, offset, &result, whence));
 #else
 #error "lseek and _llseek syscalls not available to perform a seek operation."
 #endif
@@ -82,7 +85,8 @@ ErrorOr<long> seek_func(File *f, long offset, int whence) {
 
 int close_func(File *f) {
   auto *lf = reinterpret_cast<LinuxFile *>(f);
-  int ret = __llvm_libc::syscall_impl(SYS_close, lf->get_fd());
+  int ret =
+      static_cast<int>(__llvm_libc::syscall_impl(SYS_close, lf->get_fd()));
   if (ret < 0) {
     return -ret;
   }
@@ -124,10 +128,11 @@ ErrorOr<File *> openfile(const char *path, const char *mode) {
       S_IRUSR | S_IWUSR | S_IRGRP | S_IWGRP | S_IROTH | S_IWOTH;
 
 #ifdef SYS_open
-  int fd = __llvm_libc::syscall_impl(SYS_open, path, open_flags, OPEN_MODE);
+  int fd = static_cast<int>(
+      __llvm_libc::syscall_impl(SYS_open, path, open_flags, OPEN_MODE));
 #elif defined(SYS_openat)
-  int fd = __llvm_libc::syscall_impl(SYS_openat, AT_FDCWD, path, open_flags,
-                                     OPEN_MODE);
+  int fd = static_cast<int>(__llvm_libc::syscall_impl(
+      SYS_openat, AT_FDCWD, path, open_flags, OPEN_MODE));
 #else
 #error "open and openat syscalls not available."
 #endif
