+++ libc/src/sys/sendfile/linux/sendfile.cpp
@@ -19,10 +19,10 @@ namespace __llvm_libc {
 
 LLVM_LIBC_FUNCTION(ssize_t, sendfile,
                    (int out_fd, int in_fd, off_t *offset, size_t count)) {
-  long ret =
-      __llvm_libc::syscall_impl(SYS_sendfile, in_fd, out_fd, offset, count);
+  ssize_t ret = static_cast<ssize_t>(
+      __llvm_libc::syscall_impl(SYS_sendfile, in_fd, out_fd, offset, count));
   if (ret < 0) {
-    libc_errno = -ret;
+    libc_errno = static_cast<int>(-ret);
     return -1;
   }
   return ret;
