+++ libc/src/termios/linux/tcdrain.cpp
@@ -19,7 +19,8 @@
 namespace __llvm_libc {
 
 LLVM_LIBC_FUNCTION(int, tcdrain, (int fd)) {
-  long ret = __llvm_libc::syscall_impl(SYS_ioctl, fd, TCSBRK, 1);
+  int ret =
+      static_cast<int>(__llvm_libc::syscall_impl(SYS_ioctl, fd, TCSBRK, 1));
   if (ret < 0) {
     libc_errno = -ret;
     return -1;
