+++ libc/src/time/nanosleep.cpp
@@ -19,7 +19,8 @@ namespace __llvm_libc {
 // TODO(michaelrj): Move this into time/linux with the other syscalls.
 LLVM_LIBC_FUNCTION(int, nanosleep,
                    (const struct timespec *req, struct timespec *rem)) {
-  int ret = __llvm_libc::syscall_impl(SYS_nanosleep, req, rem);
+  int ret =
+      static_cast<int>(__llvm_libc::syscall_impl(SYS_nanosleep, req, rem));
   if (ret < 0) {
     libc_errno = -ret;
     return -1;
