+++ libc/src/unistd/linux/unlinkat.cpp
@@ -19,7 +19,8 @@ namespace __llvm_libc {
 
 LLVM_LIBC_FUNCTION(int, unlinkat, (int dfd, const char *path, int flags)) {
 #ifdef SYS_unlinkat
-  long ret = __llvm_libc::syscall_impl(SYS_unlinkat, dfd, path, flags);
+  int ret = static_cast<int>(
+      __llvm_libc::syscall_impl(SYS_unlinkat, dfd, path, flags));
 #else
 #error "unlinkat syscalls not available."
 #endif
